requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
openai>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
